Make cinder a soft dependency
Add the parameter `install_cinder_client` to nova::api to make the
inclusion of the cinder::client class optional. This makes cinder a soft
dependency and the puppet-cinder module no longer has to be present in
environments where cinder is not deployed.
The default value is true, so the behaviour of the nova::api class does
not change.

Change-Id: Ibc974cb49448b86885eaf8b817489d3345901a5f
Closes-bug: #1511316